requirements.txt
# Dependencies for fetch_ads.py and fetch_keyword_planner.py
google-ads>=28.0.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
tqdm>=4.65.0
pyarrow>=10.0.0
pandas-gbq>=0.26.1
pyyaml>=6.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
